pyproject.toml
[tool.poetry]
name = "fast-api"
version = "0.1.0"
description = ""
authors = ["USiqueira"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
psycopg2-binary = "^2.9.9"
uvicorn = "^0.27.0"
SQLAlchemy = "^2.0.25"
pytest = "^8.0.0"
python-decouple = "^3.8"
httpx = "^0.26.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.7"
asyncpg = "^0.29.0"
flake8 = "^7.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
